Add error repro test for FSDP ignored modules with mixed precision (#89971)
The ignored modules are still using the original precision, which
leads to the following error.

```
RuntimeError: mat1 and mat2 must have the same dtype
```

This is not blocking me at the moment, but the fix seems not too
hard. We can add a pre-forward hook to each ignored module to
convert activations to original precision, and a post-forward hook
to convert it back to the specified precision.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/89971
Approved by: https://github.com/awgu