Use `@pytorch//` in bazel build files (#89660)
This change aims to make bazel build more embeeding-friendly.
Namely, when PyTorch is included as an external repo in another project, it is usually included like this
```
        native.local_repository(
            name = "pytorch",
            path = ...,
            repo_mapping = repo_mapping,
        )
```
Or
```
        http_archive(
            name = "pytorch",
            urls = ...
            repo_mapping = repo_mapping,
        )
```
In this case, references to `@//` would resolve to the top-level WORKSPACE that includes PyTorch.
That makes upgrades harder because we need to carry around this patch.
Note that under some edge-case circumstances even `//` resolves to the top-level `WORKSPACE`.

This change makes the embedding of the bazel build easier without compromising anything for the main repo, since the `@pytorch//` still refers to the same thing.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/89660
Approved by: https://github.com/kit1980