Dispatch the auxiliary frobenius_norm and nuclear_norm to better implementations and deprecate them (#81763)
These functions will be legacy functions. We deprecate them, but we also
take this chance to dispatch to a more efficient and consistent implementation.
Doing so should help writing a conversion rule for these to be able to
remove them once and for all

Differential Revision: [D42354776](https://our.internmc.facebook.com/intern/diff/D42354776)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81763
Approved by: https://github.com/ngimel